requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==20.3.0
bech32==1.2.0
bip32utils==0.3.post4
boltons==21.0.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
ecdsa==0.16.1
frozenlist==1.2.0
furl==2.1.3
halo==0.0.31
idna==3.3
log-symbols==0.0.14
mnemonic==0.19
multidict==5.2.0
nest-asyncio==1.5.4
orderedmultidict==1.0.1
protobuf==3.19.1
python_decouple==3.6
requests==2.26.0
six==1.16.0
spinners==0.0.24
termcolor==1.1.0
terra-sdk==2.0.6
typing_extensions==4.0.1
urllib3==1.26.7
wrapt==1.13.3
yarl==1.7.2
